[#155] Separated out the base View mustache class
